refactor: restructure listbox code
