+++ apis/nucleus/src/components/listbox/components/__tests__/list-box-highlight.spec.js
