+++ apis/nucleus/src/components/listbox/components/__tests__/list-box-search.spec.jsx
@@ -6,7 +6,7 @@ import { createTheme, ThemeProvider } from '@nebula.js/ui/theme';
 
 const InstanceContext = React.createContext();
 const [{ default: ListBoxSearch }] = aw.mock(
-  [[require.resolve('../../../contexts/InstanceContext'), () => InstanceContext]],
+  [[require.resolve('../../../../contexts/InstanceContext'), () => InstanceContext]],
   ['../ListBoxSearch']
 );
 const theme = createTheme('dark');
